# Conflicts:
#	services/authorization/src/main/java/com/processing/authorization/entities/LimitUsage.java
#	services/authorization/src/main/java/com/processing/authorization/repositories/LimitUsageRepository.java
#	services/authorization/src/main/java/com/processing/authorization/services/AuthService.java
#	services/authorization/src/test/java/com/processing/authorization/controller/AuthControllerTest.java
#	services/authorization/src/test/java/com/processing/authorization/service/AuthServiceTest.java
#	services/card-management/src/main/java/com/processing/cardmanagement/models/Card.java
#	services/card-management/src/main/java/com/processing/cardmanagement/models/CardEntity.java
#	services/card-management/src/main/java/com/processing/cardmanagement/services/CardService.java
#	services/card-management/src/main/java/com/processing/cardmanagement/services/CardServiceImpl.java
#	services/card-management/src/test/java/com/processing/cardmanagement/controllers/CardGeneratorControllerTest.java
#	services/card-management/src/test/java/com/processing/cardmanagement/services/CardGeneratorServiceTest.java
#	services/card-management/src/test/java/com/processing/cardmanagement/services/CardServiceTest.java
#	services/common/src/main/java/com/processing/common/dto/authorization/AuthorizationRequest.java
#	services/common/src/main/java/com/processing/common/dto/cardmanagement/CardModel.java
#	services/merchant-acquirer/src/main/java/com/processing/merchantacquirer/controller/dto/AcquirerFeeRequest.java
#	services/merchant-acquirer/src/main/java/com/processing/merchantacquirer/domain/entity/AcquirerFee.java
#	services/merchant-acquirer/src/main/java/com/processing/merchantacquirer/domain/factory/AuthorizationRequestFactory.java
#	services/merchant-acquirer/src/main/java/com/processing/merchantacquirer/repository/AcquirerFeeRepository.java
#	services/merchant-acquirer/src/main/java/com/processing/merchantacquirer/service/AcquirerProvider.java
#	services/merchant-acquirer/src/test/java/com/processing/merchantacquirer/service/AcquirerProviderTest.java
#	services/terminal-simulator/src/test/java/com/processing/terminalsimulator/client/GatewayClientTest.java
#	services/terminal-simulator/src/test/java/com/processing/terminalsimulator/service/TerminalSimulatorServiceTest.java

+++ services/card-management/src/main/java/com/processing/cardmanagement/models/CardEntity.java
@@ -8,6 +8,8 @@ import lombok.Setter;
 import org.hibernate.annotations.SQLRestriction;
 
 import java.time.Instant;
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
 import java.util.UUID;
 
 @Entity
@@ -45,11 +47,11 @@ public class CardEntity {
     @Column(length = 3, nullable = false)
     private String currencyCode;
 
-    private long dailyLimit;
+    private BigDecimal dailyLimit;
 
-    private long monthlyLimit;
+    private BigDecimal monthlyLimit;
 
-    private long availableBalance;
+    private BigDecimal availableBalance;
 
     @Column(length = 10, nullable = false)
     private String issuerId;
