+++ services/card-management/src/main/java/com/processing/cardmanagement/services/CardService.java
@@ -6,6 +6,8 @@ import com.processing.cardmanagement.models.CardDraft;
 import com.processing.cardmanagement.models.CardStatus;
 import jakarta.annotation.Nullable;
 
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
 import java.time.Instant;
 import java.util.List;
 
@@ -16,9 +18,9 @@ public interface CardService {
         String bin,
         String cardholderName,
         String currencyCode,
-        long dailyLimit,
-        long monthlyLimit,
-        long initialBalance
+        BigDecimal dailyLimit,
+        BigDecimal monthlyLimit,
+        BigDecimal initialBalance
     );
 
     List<Card> createCards(List<CardDraft> data);
@@ -61,9 +63,9 @@ public interface CardService {
     Card patchCard(
         String pan,
         @Nullable CardStatus status,
-        @Nullable Long dailyLimit,
-        @Nullable Long monthlyLimit,
-        @Nullable Long availableBalance
+        @Nullable BigDecimal dailyLimit,
+        @Nullable BigDecimal monthlyLimit,
+        @Nullable BigDecimal availableBalance
     );
 
     /**
@@ -107,5 +109,5 @@ public interface CardService {
      * @return измененная карта
      * @throws CardNotFoundException если карта не найдена
      */
-    Card reserve(String pan, long amount);
+    Card reserve(String pan, BigDecimal amount);
 }
