+++ services/card-management/src/test/java/com/processing/cardmanagement/services/CardGeneratorServiceTest.java
@@ -13,6 +13,8 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.time.Instant;
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
 import java.time.YearMonth;
 import java.util.List;
 import java.util.UUID;
@@ -36,10 +38,10 @@ public class CardGeneratorServiceTest {
     private MeterRegistry meterRegistry;
 
     private final CardGeneratorOptions generatorOptions = new CardGeneratorOptions(
-        1_000_000,
-        50_000_000,
-        5_000_000,
-        30_000_000,
+        BigDecimal.valueOf(1_000_000),
+        BigDecimal.valueOf(50_000_000),
+        BigDecimal.valueOf(5_000_000),
+        BigDecimal.valueOf(30_000_000),
         "643"
     );
