+++ services/merchant-acquirer/src/main/java/com/processing/merchantacquirer/domain/entity/AcquirerFee.java
@@ -8,6 +8,8 @@ import lombok.NoArgsConstructor;
 
 import java.time.Instant;
 
+import java.math.BigDecimal;
+
 
 @Entity
 @Data
@@ -19,6 +21,6 @@ public class AcquirerFee {
     private String stan;
     private String pan;
     private String terminalId;
-    private Long acquirerFee;
-    private Long amount;
+    private BigDecimal acquirerFee;
+    private BigDecimal amount;
 }
