+++ services/merchant-acquirer/src/main/java/com/processing/merchantacquirer/domain/entity/Scenario.java
@@ -1,5 +1,6 @@
 package com.processing.merchantacquirer.domain.entity;
 
+import java.math.BigDecimal;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -8,8 +9,8 @@ import lombok.Data;
 @AllArgsConstructor
 public class Scenario {
   private List<String> mcc;
-  private int countLower;
-  private int countUpper;
+  private BigDecimal countLower;
+  private BigDecimal countUpper;
   private String timeLower;
   private String timeUpper;
   private int avgApproved;
